Save tokens from Google.
